Move view metrics in dedicated files
