+++ packages/rum-core/src/domain/rumEventsCollection/view/viewMetrics/trackLoadingTime.ts
@@ -1,0 +1,46 @@
+import type { ClocksState, Duration, Observable } from '@datadog/browser-core'
+import { elapsed } from '@datadog/browser-core'
+import { waitPageActivityEnd } from '../../../waitPageActivityEnd'
+import type { RumConfiguration } from '../../../configuration'
+import type { LifeCycle } from '../../../lifeCycle'
+import { ViewLoadingType } from '../../../../rawRumEvent.types'
+
+export function trackLoadingTime(
+  lifeCycle: LifeCycle,
+  domMutationObservable: Observable<void>,
+  configuration: RumConfiguration,
+  loadType: ViewLoadingType,
+  viewStart: ClocksState,
+  callback: (loadingTime: Duration) => void
+) {
+  let isWaitingForLoadEvent = loadType === ViewLoadingType.INITIAL_LOAD
+  let isWaitingForActivityLoadingTime = true
+  const loadingTimeCandidates: Duration[] = []
+
+  function invokeCallbackIfAllCandidatesAreReceived() {
+    if (!isWaitingForActivityLoadingTime && !isWaitingForLoadEvent && loadingTimeCandidates.length > 0) {
+      callback(Math.max(...loadingTimeCandidates) as Duration)
+    }
+  }
+
+  const { stop } = waitPageActivityEnd(lifeCycle, domMutationObservable, configuration, (event) => {
+    if (isWaitingForActivityLoadingTime) {
+      isWaitingForActivityLoadingTime = false
+      if (event.hadActivity) {
+        loadingTimeCandidates.push(elapsed(viewStart.timeStamp, event.end))
+      }
+      invokeCallbackIfAllCandidatesAreReceived()
+    }
+  })
+
+  return {
+    stop,
+    setLoadEvent: (loadEvent: Duration) => {
+      if (isWaitingForLoadEvent) {
+        isWaitingForLoadEvent = false
+        loadingTimeCandidates.push(loadEvent)
+        invokeCallbackIfAllCandidatesAreReceived()
+      }
+    },
+  }
+}
